TNSR v22.10-2

+++ lib/zlog.c
@@ -47,25 +47,36 @@
 #include <mach/mach_traps.h>
 #endif
 
+#ifdef HAVE_LIBUNWIND
+#define UNW_LOCAL_ONLY
+#include <libunwind.h>
+#include <dlfcn.h>
+#endif
+
 #include "memory.h"
 #include "atomlist.h"
 #include "printfrr.h"
 #include "frrcu.h"
 #include "zlog.h"
+#include "libfrr_trace.h"
+#include "thread.h"
 
-DEFINE_MTYPE_STATIC(LIB, LOG_MESSAGE,  "log message")
-DEFINE_MTYPE_STATIC(LIB, LOG_TLSBUF,   "log thread-local buffer")
+DEFINE_MTYPE_STATIC(LIB, LOG_MESSAGE,  "log message");
+DEFINE_MTYPE_STATIC(LIB, LOG_TLSBUF,   "log thread-local buffer");
 
 DEFINE_HOOK(zlog_init, (const char *progname, const char *protoname,
 			unsigned short instance, uid_t uid, gid_t gid),
-		       (progname, protoname, instance, uid, gid))
-DEFINE_KOOH(zlog_fini, (), ())
+		       (progname, protoname, instance, uid, gid));
+DEFINE_KOOH(zlog_fini, (), ());
 DEFINE_HOOK(zlog_aux_init, (const char *prefix, int prio_min),
-			   (prefix, prio_min))
+			   (prefix, prio_min));
 
 char zlog_prefix[128];
 size_t zlog_prefixsz;
 int zlog_tmpdirfd = -1;
+int zlog_instance = -1;
+
+static atomic_bool zlog_ec = true, zlog_xid = true;
 
 /* these are kept around because logging is initialized (and directories
  * & files created) before zprivs code switches to the FRR user;  therefore
@@ -77,6 +88,11 @@ static gid_t zlog_gid = -1;
 
 DECLARE_ATOMLIST(zlog_targets, struct zlog_target, head);
 static struct zlog_targets_head zlog_targets;
+
+/* Global setting for buffered vs immediate output. The default is
+ * per-pthread buffering.
+ */
+static bool default_immediate;
 
 /* cf. zlog.h for additional comments on this struct.
  *
@@ -93,11 +109,13 @@ struct zlog_msg {
 
 	const char *fmt;
 	va_list args;
+	const struct xref_logmsg *xref;
 
 	char *stackbuf;
 	size_t stackbufsz;
 	char *text;
 	size_t textlen;
+	size_t hdrlen;
 
 	/* This is always ISO8601 with sub-second precision 9 here, it's
 	 * converted for callers as needed.  ts_dot points to the "."
@@ -107,8 +125,23 @@ struct zlog_msg {
 	 * Valid if ZLOG_TS_ISO8601 is set.
 	 * (0 if timestamp has not been formatted yet)
 	 */
-	uint32_t ts_flags;
 	char ts_str[32], *ts_dot, ts_zonetail[8];
+	uint32_t ts_flags;
+
+	/* "mmm dd hh:mm:ss" for 3164 legacy syslog - too dissimilar from
+	 * the above, so just kept separately here.
+	 */
+	uint32_t ts_3164_flags;
+	char ts_3164_str[16];
+
+	/* at the time of writing, 16 args was the actual maximum of arguments
+	 * to a single zlog call.  Particularly printing flag bitmasks seems
+	 * to drive this.  That said, the overhead of dynamically sizing this
+	 * probably outweighs the value.  If anything, a printfrr extension
+	 * for printing flag bitmasks might be a good idea.
+	 */
+	struct fmt_outpos argpos[24];
+	size_t n_argpos;
 };
 
 /* thread-local log message buffering
@@ -138,6 +171,7 @@ struct zlog_msg {
 struct zlog_tls {
 	char *mmbuf;
 	size_t bufpos;
+	bool do_unlink;
 
 	size_t nmsgs;
 	struct zlog_msg msgs[TLS_LOG_MAXMSG];
@@ -194,8 +228,15 @@ static inline void zlog_tls_set(struct zlog_tls *val)
 #endif
 
 #ifdef CAN_DO_TLS
-static long zlog_gettid(void)
+static intmax_t zlog_gettid(void)
 {
+#ifndef __OpenBSD__
+	/* accessing a TLS variable is much faster than a syscall */
+	static thread_local intmax_t cached_tid = -1;
+	if (cached_tid != -1)
+		return cached_tid;
+#endif
+
 	long rv = -1;
 #ifdef HAVE_PTHREAD_GETTHREADID_NP
 	rv = pthread_getthreadid_np();
@@ -214,6 +255,10 @@ static long zlog_gettid(void)
 #elif defined(__APPLE__)
 	rv = mach_thread_self();
 	mach_port_deallocate(mach_task_self(), rv);
+#endif
+
+#ifndef __OpenBSD__
+	cached_tid = rv;
 #endif
 	return rv;
 }
@@ -234,7 +279,7 @@ void zlog_tls_buffer_init(void)
 	for (i = 0; i < array_size(zlog_tls->msgp); i++)
 		zlog_tls->msgp[i] = &zlog_tls->msgs[i];
 
-	snprintfrr(mmpath, sizeof(mmpath), "logbuf.%ld", zlog_gettid());
+	snprintfrr(mmpath, sizeof(mmpath), "logbuf.%jd", zlog_gettid());
 
 	mmfd = openat(zlog_tmpdirfd, mmpath,
 		      O_RDWR | O_CREAT | O_EXCL | O_CLOEXEC, 0600);
@@ -262,13 +307,14 @@ void zlog_tls_buffer_init(void)
 			 mmpath, strerror(errno));
 		goto out_anon_unlink;
 	}
+	zlog_tls->do_unlink = true;
 
 	close(mmfd);
 	zlog_tls_set(zlog_tls);
 	return;
 
 out_anon_unlink:
-	unlink(mmpath);
+	unlinkat(zlog_tmpdirfd, mmpath, 0);
 	close(mmfd);
 out_anon:
 
@@ -292,14 +338,16 @@ out_anon:
 void zlog_tls_buffer_fini(void)
 {
 	char mmpath[MAXPATHLEN];
+	struct zlog_tls *zlog_tls = zlog_tls_get();
+	bool do_unlink = zlog_tls ? zlog_tls->do_unlink : false;
 
 	zlog_tls_buffer_flush();
 
-	zlog_tls_free(zlog_tls_get());
+	zlog_tls_free(zlog_tls);
 	zlog_tls_set(NULL);
 
-	snprintfrr(mmpath, sizeof(mmpath), "logbuf.%ld", zlog_gettid());
-	if (unlinkat(zlog_tmpdirfd, mmpath, 0))
+	snprintfrr(mmpath, sizeof(mmpath), "logbuf.%jd", zlog_gettid());
+	if (do_unlink && unlinkat(zlog_tmpdirfd, mmpath, 0))
 		zlog_err("unlink logbuf: %s (%d)", strerror(errno), errno);
 }
 
@@ -312,6 +360,24 @@ void zlog_tls_buffer_fini(void)
 {
 }
 #endif
+
+void zlog_msg_pid(struct zlog_msg *msg, intmax_t *pid, intmax_t *tid)
+{
+#ifndef __OpenBSD__
+	static thread_local intmax_t cached_pid = -1;
+	if (cached_pid != -1)
+		*pid = cached_pid;
+	else
+		cached_pid = *pid = (intmax_t)getpid();
+#else
+	*pid = (intmax_t)getpid();
+#endif
+#ifdef CAN_DO_TLS
+	*tid = zlog_gettid();
+#else
+	*tid = *pid;
+#endif
+}
 
 static inline void zlog_tls_free(void *arg)
 {
@@ -335,7 +401,7 @@ void zlog_tls_buffer_flush(void)
 		return;
 
 	rcu_read_lock();
-	frr_each (zlog_targets, &zlog_targets, zt) {
+	frr_each_safe (zlog_targets, &zlog_targets, zt) {
 		if (!zt->logfn)
 			continue;
 
@@ -348,12 +414,14 @@ void zlog_tls_buffer_flush(void)
 }
 
 
-static void vzlog_notls(int prio, const char *fmt, va_list ap)
+static void vzlog_notls(const struct xref_logmsg *xref, int prio,
+			const char *fmt, va_list ap)
 {
 	struct zlog_target *zt;
 	struct zlog_msg stackmsg = {
 		.prio = prio & LOG_PRIMASK,
 		.fmt = fmt,
+		.xref = xref,
 	}, *msg = &stackmsg;
 	char stackbuf[512];
 
@@ -363,7 +431,7 @@ static void vzlog_notls(int prio, const char *fmt, va_list ap)
 	msg->stackbufsz = sizeof(stackbuf);
 
 	rcu_read_lock();
-	frr_each (zlog_targets, &zlog_targets, zt) {
+	frr_each_safe (zlog_targets, &zlog_targets, zt) {
 		if (prio > zt->prio_min)
 			continue;
 		if (!zt->logfn)
@@ -378,14 +446,14 @@ static void vzlog_notls(int prio, const char *fmt, va_list ap)
 		XFREE(MTYPE_LOG_MESSAGE, msg->text);
 }
 
-static void vzlog_tls(struct zlog_tls *zlog_tls, int prio,
-		      const char *fmt, va_list ap)
+static void vzlog_tls(struct zlog_tls *zlog_tls, const struct xref_logmsg *xref,
+		      int prio, const char *fmt, va_list ap)
 {
 	struct zlog_target *zt;
 	struct zlog_msg *msg;
 	char *buf;
 	bool ignoremsg = true;
-	bool immediate = false;
+	bool immediate = default_immediate;
 
 	/* avoid further processing cost if no target wants this message */
 	rcu_read_lock();
@@ -412,6 +480,7 @@ static void vzlog_tls(struct zlog_tls *zlog_tls, int prio,
 	msg->stackbufsz = TLS_LOG_BUF_SIZE - zlog_tls->bufpos - 1;
 	msg->fmt = fmt;
 	msg->prio = prio & LOG_PRIMASK;
+	msg->xref = xref;
 	if (msg->prio < LOG_INFO)
 		immediate = true;
 
@@ -446,14 +515,133 @@ static void vzlog_tls(struct zlog_tls *zlog_tls, int prio,
 		XFREE(MTYPE_LOG_MESSAGE, msg->text);
 }
 
-void vzlog(int prio, const char *fmt, va_list ap)
+static void zlog_backtrace_msg(const struct xref_logmsg *xref, int prio)
+{
+	struct thread *tc = pthread_getspecific(thread_current);
+	const char *uid = xref->xref.xrefdata->uid;
+	bool found_thread = false;
+
+	zlog(prio, "| (%s) message in thread %jd, at %s(), %s:%d", uid,
+	     zlog_gettid(), xref->xref.func, xref->xref.file, xref->xref.line);
+
+#ifdef HAVE_LIBUNWIND
+	const char *threadfunc = tc ? tc->xref->funcname : NULL;
+	bool found_caller = false;
+	unw_cursor_t cursor;
+	unw_context_t uc;
+	unw_word_t ip, off, sp;
+	Dl_info dlinfo;
+
+	unw_getcontext(&uc);
+
+	unw_init_local(&cursor, &uc);
+	while (unw_step(&cursor) > 0) {
+		char buf[96], name[128] = "?";
+		bool is_thread = false;
+
+		unw_get_reg(&cursor, UNW_REG_IP, &ip);
+		unw_get_reg(&cursor, UNW_REG_SP, &sp);
+
+		if (unw_is_signal_frame(&cursor))
+			zlog(prio, "| (%s)    ---- signal ----", uid);
+
+		if (!unw_get_proc_name(&cursor, buf, sizeof(buf), &off)) {
+			if (!strcmp(buf, xref->xref.func))
+				found_caller = true;
+			if (threadfunc && !strcmp(buf, threadfunc))
+				found_thread = is_thread = true;
+
+			snprintf(name, sizeof(name), "%s+%#lx", buf, (long)off);
+		}
+
+		if (!found_caller)
+			continue;
+
+		if (dladdr((void *)ip, &dlinfo))
+			zlog(prio, "| (%s) %-36s %16lx+%08lx %16lx %s", uid,
+			     name, (long)dlinfo.dli_fbase,
+			     (long)ip - (long)dlinfo.dli_fbase, (long)sp,
+			     dlinfo.dli_fname);
+		else
+			zlog(prio, "| (%s) %-36s %16lx %16lx", uid, name,
+			     (long)ip, (long)sp);
+
+		if (is_thread)
+			zlog(prio, "| (%s) ^- scheduled from %s(), %s:%u", uid,
+			     tc->xref->xref.func, tc->xref->xref.file,
+			     tc->xref->xref.line);
+	}
+#elif defined(HAVE_GLIBC_BACKTRACE)
+	void *frames[64];
+	char **names = NULL;
+	int n_frames, i;
+
+	n_frames = backtrace(frames, array_size(frames));
+	if (n_frames < 0)
+		n_frames = 0;
+	if (n_frames)
+		names = backtrace_symbols(frames, n_frames);
+
+	for (i = 0; i < n_frames; i++) {
+		void *retaddr = frames[i];
+		char *loc = names[i];
+
+		zlog(prio, "| (%s) %16lx %-36s", uid, (long)retaddr, loc);
+	}
+	free(names);
+#endif
+	if (!found_thread && tc)
+		zlog(prio, "| (%s) scheduled from %s(), %s:%u", uid,
+		     tc->xref->xref.func, tc->xref->xref.file,
+		     tc->xref->xref.line);
+}
+
+void vzlogx(const struct xref_logmsg *xref, int prio,
+	    const char *fmt, va_list ap)
 {
 	struct zlog_tls *zlog_tls = zlog_tls_get();
 
+#ifdef HAVE_LTTNG
+	va_list copy;
+	va_copy(copy, ap);
+	char *msg = vasprintfrr(MTYPE_LOG_MESSAGE, fmt, copy);
+
+	switch (prio) {
+	case LOG_ERR:
+		frrtracelog(TRACE_ERR, msg);
+		break;
+	case LOG_WARNING:
+		frrtracelog(TRACE_WARNING, msg);
+		break;
+	case LOG_DEBUG:
+		frrtracelog(TRACE_DEBUG, msg);
+		break;
+	case LOG_NOTICE:
+		frrtracelog(TRACE_DEBUG, msg);
+		break;
+	case LOG_INFO:
+	default:
+		frrtracelog(TRACE_INFO, msg);
+		break;
+	}
+
+	va_end(copy);
+	XFREE(MTYPE_LOG_MESSAGE, msg);
+#endif
+
 	if (zlog_tls)
-		vzlog_tls(zlog_tls, prio, fmt, ap);
+		vzlog_tls(zlog_tls, xref, prio, fmt, ap);
 	else
-		vzlog_notls(prio, fmt, ap);
+		vzlog_notls(xref, prio, fmt, ap);
+
+	if (xref) {
+		struct xrefdata_logmsg *xrdl;
+
+		xrdl = container_of(xref->xref.xrefdata, struct xrefdata_logmsg,
+				    xrefdata);
+		if (xrdl->fl_print_bt)
+			zlog_backtrace_msg(xref, prio);
+	}
 }
 
 void zlog_sigsafe(const char *text, size_t len)
@@ -481,34 +669,132 @@ void zlog_sigsafe(const char *text, size_t len)
 	}
 }
 
+void _zlog_assert_failed(const struct xref_assert *xref, const char *extra, ...)
+{
+	va_list ap;
+	static bool assert_in_assert; /* "global-ish" variable, init to 0 */
+
+	if (assert_in_assert)
+		abort();
+	assert_in_assert = true;
+
+	if (extra) {
+		struct va_format vaf;
+
+		va_start(ap, extra);
+		vaf.fmt = extra;
+		vaf.va = &ap;
+
+		zlog(LOG_CRIT,
+		     "%s:%d: %s(): assertion (%s) failed, extra info: %pVA",
+		     xref->xref.file, xref->xref.line, xref->xref.func,
+		     xref->expr, &vaf);
+
+		va_end(ap);
+	} else
+		zlog(LOG_CRIT, "%s:%d: %s(): assertion (%s) failed",
+		     xref->xref.file, xref->xref.line, xref->xref.func,
+		     xref->expr);
+
+	/* abort() prints backtrace & memstats in SIGABRT handler */
+	abort();
+}
 
 int zlog_msg_prio(struct zlog_msg *msg)
 {
 	return msg->prio;
 }
 
+const struct xref_logmsg *zlog_msg_xref(struct zlog_msg *msg)
+{
+	return msg->xref;
+}
+
 const char *zlog_msg_text(struct zlog_msg *msg, size_t *textlen)
 {
 	if (!msg->text) {
 		va_list args;
+		bool do_xid, do_ec;
+		size_t need = 0, hdrlen;
+		struct fbuf fb = {
+			.buf = msg->stackbuf,
+			.pos = msg->stackbuf,
+			.len = msg->stackbufsz,
+		};
+
+		do_ec = atomic_load_explicit(&zlog_ec, memory_order_relaxed);
+		do_xid = atomic_load_explicit(&zlog_xid, memory_order_relaxed);
+
+		if (msg->xref && do_xid && msg->xref->xref.xrefdata->uid[0]) {
+			need += bputch(&fb, '[');
+			need += bputs(&fb, msg->xref->xref.xrefdata->uid);
+			need += bputch(&fb, ']');
+		}
+		if (msg->xref && do_ec && msg->xref->ec)
+			need += bprintfrr(&fb, "[EC %u]", msg->xref->ec);
+		if (need)
+			need += bputch(&fb, ' ');
+
+		msg->hdrlen = hdrlen = need;
+		assert(hdrlen < msg->stackbufsz);
+
+		fb.outpos = msg->argpos;
+		fb.outpos_n = array_size(msg->argpos);
+		fb.outpos_i = 0;
 
 		va_copy(args, msg->args);
-		msg->text = vasnprintfrr(MTYPE_LOG_MESSAGE, msg->stackbuf,
-					 msg->stackbufsz, msg->fmt, args);
-		msg->textlen = strlen(msg->text);
+		need += vbprintfrr(&fb, msg->fmt, args);
 		va_end(args);
+
+		msg->textlen = need;
+		need += bputch(&fb, '\n');
+
+		if (need <= msg->stackbufsz)
+			msg->text = msg->stackbuf;
+		else {
+			msg->text = XMALLOC(MTYPE_LOG_MESSAGE, need);
+
+			memcpy(msg->text, msg->stackbuf, hdrlen);
+
+			fb.buf = msg->text;
+			fb.len = need;
+			fb.pos = msg->text + hdrlen;
+			fb.outpos_i = 0;
+
+			va_copy(args, msg->args);
+			vbprintfrr(&fb, msg->fmt, args);
+			va_end(args);
+
+			bputch(&fb, '\n');
+		}
+
+		msg->n_argpos = fb.outpos_i;
 	}
 	if (textlen)
 		*textlen = msg->textlen;
 	return msg->text;
 }
 
+void zlog_msg_args(struct zlog_msg *msg, size_t *hdrlen, size_t *n_argpos,
+		   const struct fmt_outpos **argpos)
+{
+	if (!msg->text)
+		zlog_msg_text(msg, NULL);
+
+	if (hdrlen)
+		*hdrlen = msg->hdrlen;
+	if (n_argpos)
+		*n_argpos = msg->n_argpos;
+	if (argpos)
+		*argpos = msg->argpos;
+}
+
 #define ZLOG_TS_FORMAT		(ZLOG_TS_ISO8601 | ZLOG_TS_LEGACY)
 #define ZLOG_TS_FLAGS		~ZLOG_TS_PREC
 
-size_t zlog_msg_ts(struct zlog_msg *msg, char *out, size_t outsz,
-		   uint32_t flags)
+size_t zlog_msg_ts(struct zlog_msg *msg, struct fbuf *out, uint32_t flags)
 {
+	size_t outsz = out ? (out->buf + out->len - out->pos) : 0;
 	size_t len1;
 
 	if (!(flags & ZLOG_TS_FORMAT))
@@ -550,34 +836,101 @@ size_t zlog_msg_ts(struct zlog_msg *msg, char *out, size_t outsz,
 		len1 = strlen(msg->ts_str);
 
 	if (flags & ZLOG_TS_LEGACY) {
-		if (len1 + 1 > outsz)
-			return 0;
+		if (!out)
+			return len1;
+
+		if (len1 > outsz) {
+			memset(out->pos, 0, outsz);
+			out->pos += outsz;
+			return len1;
+		}
 
 		/* just swap out the formatting, faster than redoing it */
 		for (char *p = msg->ts_str; p < msg->ts_str + len1; p++) {
 			switch (*p) {
 			case '-':
-				*out++ = '/';
+				*out->pos++ = '/';
 				break;
 			case 'T':
-				*out++ = ' ';
+				*out->pos++ = ' ';
 				break;
 			default:
-				*out++ = *p;
+				*out->pos++ = *p;
 			}
 		}
-		*out = '\0';
 		return len1;
 	} else {
 		size_t len2 = strlen(msg->ts_zonetail);
 
-		if (len1 + len2 + 1 > outsz)
-			return 0;
-		memcpy(out, msg->ts_str, len1);
-		memcpy(out + len1, msg->ts_zonetail, len2);
-		out[len1 + len2] = '\0';
+		if (!out)
+			return len1 + len2;
+
+		if (len1 + len2 > outsz) {
+			memset(out->pos, 0, outsz);
+			out->pos += outsz;
+			return len1 + len2;
+		}
+
+		memcpy(out->pos, msg->ts_str, len1);
+		out->pos += len1;
+		memcpy(out->pos, msg->ts_zonetail, len2);
+		out->pos += len2;
 		return len1 + len2;
 	}
+}
+
+size_t zlog_msg_ts_3164(struct zlog_msg *msg, struct fbuf *out, uint32_t flags)
+{
+	flags &= ZLOG_TS_UTC;
+
+	if (!msg->ts_3164_str[0] || flags != msg->ts_3164_flags) {
+		/* these are "hardcoded" in RFC3164, so they're here too... */
+		static const char *const months[12] = {
+			"Jan", "Feb", "Mar", "Apr", "May", "Jun",
+			"Jul", "Aug", "Sep", "Oct", "Nov", "Dec",
+		};
+		struct tm tm;
+
+		/* RFC3164 explicitly asks for local time, but common usage
+		 * also includes UTC.
+		 */
+		if (flags & ZLOG_TS_UTC)
+			gmtime_r(&msg->ts.tv_sec, &tm);
+		else
+			localtime_r(&msg->ts.tv_sec, &tm);
+
+		snprintfrr(msg->ts_3164_str, sizeof(msg->ts_3164_str),
+			   "%3s %2d %02d:%02d:%02d", months[tm.tm_mon],
+			   tm.tm_mday, tm.tm_hour, tm.tm_min, tm.tm_sec);
+
+		msg->ts_3164_flags = flags;
+	}
+	return bputs(out, msg->ts_3164_str);
+}
+
+void zlog_msg_tsraw(struct zlog_msg *msg, struct timespec *ts)
+{
+	memcpy(ts, &msg->ts, sizeof(*ts));
+}
+
+void zlog_set_prefix_ec(bool enable)
+{
+	atomic_store_explicit(&zlog_ec, enable, memory_order_relaxed);
+}
+
+bool zlog_get_prefix_ec(void)
+{
+	return atomic_load_explicit(&zlog_ec, memory_order_relaxed);
+}
+
+void zlog_set_prefix_xid(bool enable)
+{
+	atomic_store_explicit(&zlog_xid, enable, memory_order_relaxed);
+}
+
+bool zlog_get_prefix_xid(void)
+{
+	return atomic_load_explicit(&zlog_xid, memory_order_relaxed);
 }
 
 /* setup functions */
@@ -607,6 +960,14 @@ struct zlog_target *zlog_target_replace(struct zlog_target *oldzt,
 	return oldzt;
 }
 
+/*
+ * Enable or disable 'immediate' output - default is to buffer
+ * each pthread's messages.
+ */
+void zlog_set_immediate(bool set_p)
+{
+	default_immediate = set_p;
+}
 
 /* common init */
 
@@ -627,6 +988,7 @@ void zlog_init(const char *progname, const char *protoname,
 {
 	zlog_uid = uid;
 	zlog_gid = gid;
+	zlog_instance = instance;
 
 	if (instance) {
 		snprintfrr(zlog_tmpdir, sizeof(zlog_tmpdir),
